general bugfixing and improvements
